+++ trino/versions.py
@@ -13,4 +13,11 @@ versions = [
         "jmx_exporter": "1.2.0",
         "trino-storage-connector": "470",
     },
+    {
+        "product": "476",
+        "java-base": "24",
+        "java-devel": "24",
+        "jmx_exporter": "1.2.0",
+        "trino-storage-connector": "476",
+    },
 ]